Added options for user class

+++ console.py
@@ -7,6 +7,7 @@ import cmd
 from models.base_model import BaseModel
 from models.engine.file_storage import FileStorage
 from models.__init__ import storage
+from models.user import User
 import json
 
 
@@ -14,15 +15,16 @@ class HBNBCommand(cmd.Cmd):
     intro = ""
     prompt = "(hbnb) "
     file = None
-    models = ["BaseModel"]
+    models = ["BaseModel", "User"]
     all_instances = []
-    base_model_instances = []
+    User_instances = []
+    BaseModel_instances = []
     try:
 
         storage.reload()
         objs = storage.all()
         for key, val in objs.items():
-            base_model_instances.append(val)
+            all_instances.append(val)
     except Exception as Err:
         print(Err)
 
@@ -38,7 +40,7 @@ class HBNBCommand(cmd.Cmd):
         if len(list) == 1:
             print('** instance id missing **')
             return False
-        ids = [el.id for el in self.base_model_instances]
+        ids = [el.id for el in self.all_instances]
         if list[1] not in ids:
             print('** no instance found **')
             return False
@@ -69,7 +71,7 @@ class HBNBCommand(cmd.Cmd):
                 bm = eval(f"{args[0]}()")
                 print(bm.id)
                 bm.save()
-                self.base_model_instances.append(bm)
+                self.all_instances.append(bm)
             except Exception as Err:
                 print(Err)
 
@@ -79,8 +81,8 @@ class HBNBCommand(cmd.Cmd):
         list = args.split()
         if self.check_arguments(args) is False:
             return ''
-        for ins in self.base_model_instances:
-            if ins.id == list[1]:
+        for ins in self.all_instances:
+            if ins.id == list[1] and list[0] == ins.__class__.__name__:
                 print(ins)
 
     def do_destroy(self, args=''):
@@ -88,24 +90,26 @@ class HBNBCommand(cmd.Cmd):
         list = args.split()
         if self.check_arguments(args) is False:
             return ''
-        for el in self.base_model_instances:
-            if el.id == list[1]:
-                index = self.base_model_instances.index(el)
-                self.base_model_instances.pop(index)
+        for el in self.all_instances:
+            if el.id == list[1] and list[0] == el.__class__.__name__:
+                index = self.all_instances.index(el)
+                self.all_instances.pop(index)
                 storage.destroy(el)
 
     def do_all(self, args=''):
         'Show all strs of base_model_instances of a model'
-        list = args.split()
-        if len(list) == 0:
-            for ins in self.base_model_instances:
+
+        if len(args) == 0:
+            for ins in self.all_instances:
                 print(ins)
                 return ''
-        if list[0] not in self.models:
+        if args not in self.models:
             print('** class doesn\'t exist **')
             return ''
-        for ins in self.base_model_instances:
-            print(ins)
+        for ins in self.all_instances:
+            if ins.__class__.__name__ == args:
+                print(f'{ins.__class__.__name__} === {args}')
+                print(ins)
 
     def do_update(self, args=''):
         'Updates an instance'
@@ -119,8 +123,8 @@ class HBNBCommand(cmd.Cmd):
             print('** value missing **')
             return ''
         try:
-            for ins in self.base_model_instances:
-                if ins.id == list[1]:
+            for ins in self.all_instances:
+                if ins.id == list[1] and ins.__class__.__name__ == list[0]:
                     setattr(ins, list[2], list[3])
         except Exception as Err:
             print(Err)
